Update cartItems array when user clicks add to bag button
